barcode_tool/requirements.txt
streamlit==1.29.0
python-barcode==0.16.1
pillow==9.5.0
img2pdf==0.4.0